add authentication certificate in test resources because on rebuild every time a new Rsa key was generated with a new keypair
