+++ src/main/java/com/ozius/internship/project/security/WebConfig.java
@@ -1,11 +1,19 @@
 package com.ozius.internship.project.security;
 
+import com.ozius.internship.project.security.jwt.FusedClaimConverter;
+import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
+@EnableMethodSecurity(jsr250Enabled = true, securedEnabled = true) //this annotation can be put on any config class but this is the best place
 public class WebConfig {
 
     @Bean
@@ -18,6 +26,32 @@ public class WebConfig {
                         .allowedOrigins("http://localhost:3000");
             }
         };
+    }
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http.authorizeHttpRequests(auth -> auth
+                .requestMatchers("/authenticate").permitAll()
+                .requestMatchers("/register-client").permitAll()
+                .requestMatchers("/users/{email}").permitAll()
+                .requestMatchers("/images/**").permitAll()
+                .requestMatchers("/categories/**").permitAll()
+                .requestMatchers("/products/**").permitAll()
+                .requestMatchers("/error").permitAll()
+                .requestMatchers(HttpMethod.OPTIONS, "/**").permitAll()
+                .requestMatchers(PathRequest.toH2Console()).permitAll()
+                .anyRequest().authenticated());
+
+        http.sessionManagement(
+                session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+        );
+
+        http.httpBasic().disable();
+        http.csrf().disable();
+        http.headers().frameOptions().sameOrigin();
+        http.oauth2ResourceServer(oauth -> oauth.jwt(jwt -> jwt.jwtAuthenticationConverter(new FusedClaimConverter())));
+
+        return http.build();
     }
 
 }
